Adding owner role to questions

+++ src/api/src/services/question-service.ts
@@ -10,9 +10,15 @@ export class QuestionService implements GenericService<Question> {
     for (let q of questions) {
       let moderators = await db(DB_USER_QUESTION_TABLE)
         .withSchema(DB_SCHEMA)
-        .where({ QUESTION_ID: q.ID })
+        .where({ QUESTION_ID: q.ID, ROLE: "Moderator" })
         .select("EMAIL");
       q.moderators = moderators.map((m) => m.EMAIL);
+
+      let owners = await db(DB_USER_QUESTION_TABLE)
+        .withSchema(DB_SCHEMA)
+        .where({ QUESTION_ID: q.ID, ROLE: "Owner" })
+        .select("EMAIL");
+      q.owners = owners.map((m) => m.EMAIL);
     }
 
     return questions;
@@ -45,10 +51,20 @@ export class QuestionService implements GenericService<Question> {
   }
 
   async setModerators(QUESTION_ID: number, emails: string[]): Promise<any> {
-    await db(DB_USER_QUESTION_TABLE).withSchema(DB_SCHEMA).where({ QUESTION_ID }).delete();
+    await db(DB_USER_QUESTION_TABLE).withSchema(DB_SCHEMA).where({ QUESTION_ID, ROLE: "Moderator" }).delete();
 
     let inserts = emails.map((e) => {
       return { EMAIL: e, ROLE: "Moderator", QUESTION_ID };
+    });
+
+    if (inserts.length > 0) return db(DB_USER_QUESTION_TABLE).withSchema(DB_SCHEMA).insert(inserts);
+  }
+
+  async setOwners(QUESTION_ID: number, emails: string[]): Promise<any> {
+    await db(DB_USER_QUESTION_TABLE).withSchema(DB_SCHEMA).where({ QUESTION_ID, ROLE: "Owner" }).delete();
+
+    let inserts = emails.map((e) => {
+      return { EMAIL: e, ROLE: "Owner", QUESTION_ID };
     });
 
     if (inserts.length > 0) return db(DB_USER_QUESTION_TABLE).withSchema(DB_SCHEMA).insert(inserts);
@@ -59,6 +75,8 @@ export class QuestionService implements GenericService<Question> {
   }
 
   async createEvent(QUESTION_ID: number, ACTION: string, DESCRIPTION: string): Promise<any> {
-    return db("QUESTION_HISTORY").withSchema(DB_SCHEMA).insert({ QUESTION_ID, ACTION, DESCRIPTION, CREATE_DATE: new Date() });
+    return db("QUESTION_HISTORY")
+      .withSchema(DB_SCHEMA)
+      .insert({ QUESTION_ID, ACTION, DESCRIPTION, CREATE_DATE: new Date() });
   }
 }
